feat(core): Extract ExternalTokenVerifierService from TokenExchangeService (no-changelog)
Third step of the token-exchange/identity-substrate split (IAM-1181).
TokenExchangeService bundled the verify pipeline (verifyToken,
verifyExternalToken, verifyInboundToken) - the primitive
InboundClaimVerificationHook calls on every context establishment via
ExternalTokenVerifierProxy - together with the RFC 8693 issuance flow
(embedLogin, exchange).

Extracts the verify pipeline into
identity-substrate/services/external-token-verifier.service.ts. The
consumer's TokenExchangeService now only implements embedLogin() and
exchange(), injecting ExternalTokenVerifierService from the substrate.
token-exchange.module.ts registers ExternalTokenVerifierService (not
TokenExchangeService) as the ExternalTokenVerifierProxy provider.

Still no module-boundary change - the registration happens inside
token-exchange's main-only init(), so this stays behaviorally a no-op.
