+++ packages/cli/src/modules/token-exchange/services/__tests__/token-exchange.service.test.ts
@@ -1,38 +1,24 @@
-import type { Logger } from '@n8n/backend-common';
 import { GLOBAL_MEMBER_ROLE, type User } from '@n8n/db';
-import jwt from 'jsonwebtoken';
-import { generateKeyPairSync } from 'node:crypto';
 import { mock } from 'vitest-mock-extended';
 
-import { AuthError } from '@/errors/response-errors/auth.error';
-import { BadRequestError } from '@/errors/response-errors/bad-request.error';
-import type { JtiStoreService } from '@/modules/identity-substrate/services/jti-store.service';
-import type { TrustedKeyService } from '@/modules/identity-substrate/services/trusted-key.service';
+import type { ExternalTokenVerifierService } from '@/modules/identity-substrate/services/external-token-verifier.service';
 import type { JwtService } from '@/services/jwt.service';
 
-import type { InboundAudienceService } from '../../context-establishment-hooks/inbound-audience.service';
 import type { TokenExchangeConfig } from '../../token-exchange.config';
-import { TokenExchangeAuthError } from '../../token-exchange.errors';
 import type { ResolvedTrustedKey } from '../../token-exchange.schemas';
 import type { IdentityResolutionService } from '../identity-resolution.service';
 import { TokenExchangeService } from '../token-exchange.service';
 
-const logger = mock<Logger>({ scoped: vi.fn().mockReturnThis() });
-const trustedKeyStore = mock<TrustedKeyService>();
-const jtiStore = mock<JtiStoreService>();
+const externalTokenVerifierService = mock<ExternalTokenVerifierService>();
 const identityResolutionService = mock<IdentityResolutionService>();
 const config = mock<TokenExchangeConfig>();
 const jwtService = mock<JwtService>();
-const inboundAudienceService = mock<InboundAudienceService>();
 
 const service = new TokenExchangeService(
-	logger,
-	trustedKeyStore,
-	jtiStore,
+	externalTokenVerifierService,
 	identityResolutionService,
 	config,
 	jwtService,
-	inboundAudienceService,
 );
 
 const resolvedKey: ResolvedTrustedKey = {
@@ -66,21 +52,14 @@ const validClaims = {
 describe('TokenExchangeService', () => {
 	beforeEach(() => {
 		vi.clearAllMocks();
-		vi.restoreAllMocks();
 	});
 
 	describe('embedLogin', () => {
 		it('should return user on valid token', async () => {
-			vi.spyOn(jwt, 'decode').mockReturnValue({
-				header: { alg: 'RS256', kid: 'test-kid' },
-				payload: validClaims,
-				signature: 'sig',
-			} as unknown as ReturnType<typeof jwt.decode>);
-			vi.spyOn(jwt, 'verify').mockReturnValue(
-				validClaims as unknown as ReturnType<typeof jwt.verify>,
-			);
-			trustedKeyStore.getByKidAndIss.mockResolvedValue(resolvedKey);
-			jtiStore.consume.mockResolvedValue(true);
+			externalTokenVerifierService.verifyToken.mockResolvedValue({
+				claims: validClaims,
+				resolvedKey,
+			});
 			identityResolutionService.resolve.mockResolvedValue(mockUser);
 
 			const result = await service.embedLogin('valid-token');
@@ -91,14 +70,9 @@ describe('TokenExchangeService', () => {
 				issuer: 'https://issuer.example.com',
 				kid: 'test-kid',
 			});
-			expect(trustedKeyStore.getByKidAndIss).toHaveBeenCalledWith(
-				'test-kid',
-				'https://issuer.example.com',
-			);
-			expect(jtiStore.consume).toHaveBeenCalledWith(
-				'unique-jti-1',
-				new Date(validClaims.exp * 1000),
-			);
+			expect(externalTokenVerifierService.verifyToken).toHaveBeenCalledWith('valid-token', {
+				maxLifetimeSeconds: 60,
+			});
 			expect(identityResolutionService.resolve).toHaveBeenCalledWith(
 				validClaims,
 				resolvedKey.allowedRoles,
@@ -111,396 +85,70 @@ describe('TokenExchangeService', () => {
 			);
 		});
 
-		it('should throw when token cannot be decoded', async () => {
-			vi.spyOn(jwt, 'decode').mockReturnValue(null);
+		it('should propagate verification errors from the substrate verifier', async () => {
+			externalTokenVerifierService.verifyToken.mockRejectedValue(new Error('Unknown key id'));
 
-			await expect(service.embedLogin('garbage')).rejects.toThrow(BadRequestError);
-		});
-
-		it('should throw when kid is missing from JWT header', async () => {
-			vi.spyOn(jwt, 'decode').mockReturnValue({
-				header: { alg: 'RS256' },
-				payload: validClaims,
-				signature: 'sig',
-			} as unknown as ReturnType<typeof jwt.decode>);
-
-			await expect(service.embedLogin('no-kid-token')).rejects.toThrow(BadRequestError);
-		});
-
-		it('should throw when iss is missing from JWT payload', async () => {
-			const { iss: _, ...claimsWithoutIss } = validClaims;
-			vi.spyOn(jwt, 'decode').mockReturnValue({
-				header: { alg: 'RS256', kid: 'test-kid' },
-				payload: claimsWithoutIss,
-				signature: 'sig',
-			} as unknown as ReturnType<typeof jwt.decode>);
-
-			await expect(service.embedLogin('no-iss-token')).rejects.toThrow(BadRequestError);
-		});
-
-		it('should throw when kid is unknown', async () => {
-			vi.spyOn(jwt, 'decode').mockReturnValue({
-				header: { alg: 'RS256', kid: 'unknown-kid' },
-				payload: validClaims,
-				signature: 'sig',
-			} as unknown as ReturnType<typeof jwt.decode>);
-			trustedKeyStore.getByKidAndIss.mockResolvedValue(undefined);
-
-			await expect(service.embedLogin('unknown-kid-token')).rejects.toThrow(AuthError);
-		});
-
-		it('should throw when jwt.verify returns unexpected payload format', async () => {
-			vi.spyOn(jwt, 'decode').mockReturnValue({
-				header: { alg: 'RS256', kid: 'test-kid' },
-				payload: validClaims,
-				signature: 'sig',
-			} as unknown as ReturnType<typeof jwt.decode>);
-			vi.spyOn(jwt, 'verify').mockReturnValue(
-				'string-payload' as unknown as ReturnType<typeof jwt.verify>,
-			);
-			trustedKeyStore.getByKidAndIss.mockResolvedValue(resolvedKey);
-
-			await expect(service.embedLogin('string-payload-token')).rejects.toThrow(AuthError);
-		});
-
-		it('should throw when JWT signature verification fails', async () => {
-			vi.spyOn(jwt, 'decode').mockReturnValue({
-				header: { alg: 'RS256', kid: 'test-kid' },
-				payload: validClaims,
-				signature: 'sig',
-			} as unknown as ReturnType<typeof jwt.decode>);
-			vi.spyOn(jwt, 'verify').mockImplementation(() => {
-				throw new Error('invalid signature');
-			});
-			trustedKeyStore.getByKidAndIss.mockResolvedValue(resolvedKey);
-
-			await expect(service.embedLogin('bad-sig-token')).rejects.toThrow(AuthError);
-		});
-
-		it('should throw when claims fail Zod validation', async () => {
-			const invalidClaims = { ...validClaims, sub: '' };
-			vi.spyOn(jwt, 'decode').mockReturnValue({
-				header: { alg: 'RS256', kid: 'test-kid' },
-				payload: invalidClaims,
-				signature: 'sig',
-			} as unknown as ReturnType<typeof jwt.decode>);
-			vi.spyOn(jwt, 'verify').mockReturnValue(
-				invalidClaims as unknown as ReturnType<typeof jwt.verify>,
-			);
-			trustedKeyStore.getByKidAndIss.mockResolvedValue(resolvedKey);
-
-			await expect(service.embedLogin('invalid-claims-token')).rejects.toThrow();
-		});
-
-		it('should throw when token lifetime exceeds 60 seconds', async () => {
-			const longLivedClaims = { ...validClaims, iat: now, exp: now + 120 };
-			vi.spyOn(jwt, 'decode').mockReturnValue({
-				header: { alg: 'RS256', kid: 'test-kid' },
-				payload: longLivedClaims,
-				signature: 'sig',
-			} as unknown as ReturnType<typeof jwt.decode>);
-			vi.spyOn(jwt, 'verify').mockReturnValue(
-				longLivedClaims as unknown as ReturnType<typeof jwt.verify>,
-			);
-			trustedKeyStore.getByKidAndIss.mockResolvedValue(resolvedKey);
-
-			await expect(service.embedLogin('long-lived-token')).rejects.toThrow(AuthError);
-		});
-
-		it('should throw when JTI has already been consumed (replay)', async () => {
-			vi.spyOn(jwt, 'decode').mockReturnValue({
-				header: { alg: 'RS256', kid: 'test-kid' },
-				payload: validClaims,
-				signature: 'sig',
-			} as unknown as ReturnType<typeof jwt.decode>);
-			vi.spyOn(jwt, 'verify').mockReturnValue(
-				validClaims as unknown as ReturnType<typeof jwt.verify>,
-			);
-			trustedKeyStore.getByKidAndIss.mockResolvedValue(resolvedKey);
-			jtiStore.consume.mockResolvedValue(false);
-
-			await expect(service.embedLogin('replayed-token')).rejects.toThrow(AuthError);
+			await expect(service.embedLogin('bad-token')).rejects.toThrow('Unknown key id');
+			expect(identityResolutionService.resolve).not.toHaveBeenCalled();
 		});
 
 		it('should propagate error from IdentityResolutionService', async () => {
-			vi.spyOn(jwt, 'decode').mockReturnValue({
-				header: { alg: 'RS256', kid: 'test-kid' },
-				payload: validClaims,
-				signature: 'sig',
-			} as unknown as ReturnType<typeof jwt.decode>);
-			vi.spyOn(jwt, 'verify').mockReturnValue(
-				validClaims as unknown as ReturnType<typeof jwt.verify>,
-			);
-			trustedKeyStore.getByKidAndIss.mockResolvedValue(resolvedKey);
-			jtiStore.consume.mockResolvedValue(true);
+			externalTokenVerifierService.verifyToken.mockResolvedValue({
+				claims: validClaims,
+				resolvedKey,
+			});
 			identityResolutionService.resolve.mockRejectedValue(new Error('User not found'));
 
 			await expect(service.embedLogin('token')).rejects.toThrow('User not found');
 		});
 	});
 
-	describe('verifyToken', () => {
-		const mockValidToken = () => {
-			vi.spyOn(jwt, 'decode').mockReturnValue({
-				header: { alg: 'RS256', kid: 'test-kid' },
-				payload: validClaims,
-				signature: 'sig',
-			} as unknown as ReturnType<typeof jwt.decode>);
-			vi.spyOn(jwt, 'verify').mockReturnValue(
-				validClaims as unknown as ReturnType<typeof jwt.verify>,
-			);
-			trustedKeyStore.getByKidAndIss.mockResolvedValue(resolvedKey);
-		};
+	describe('exchange', () => {
+		it('should issue a token for a valid subject_token', async () => {
+			externalTokenVerifierService.verifyToken.mockResolvedValue({
+				claims: validClaims,
+				resolvedKey,
+			});
+			identityResolutionService.resolve.mockResolvedValue(mockUser);
+			config.maxTokenTtl = 900;
+			jwtService.sign.mockReturnValue('issued.jwt.token');
 
-		it('defaults consume the jti and honour the source-level audience', async () => {
-			mockValidToken();
-			jtiStore.consume.mockResolvedValue(true);
+			const result = await service.exchange({
+				grant_type: 'urn:ietf:params:oauth:grant-type:token-exchange',
+				subject_token: 'subject-token',
+			});
 
-			await service.verifyToken('valid-token');
-
-			expect(jwt.verify).toHaveBeenCalledWith(
-				'valid-token',
-				resolvedKey.key,
-				expect.objectContaining({ audience: resolvedKey.expectedAudience }),
-			);
-			expect(jtiStore.consume).toHaveBeenCalledWith(
-				validClaims.jti,
-				new Date(validClaims.exp * 1000),
-			);
+			expect(result).toMatchObject({
+				accessToken: 'issued.jwt.token',
+				subjectUserId: mockUser.id,
+				subject: validClaims.sub,
+				issuer: validClaims.iss,
+			});
+			expect(externalTokenVerifierService.verifyToken).toHaveBeenCalledWith('subject-token');
+			expect(externalTokenVerifierService.verifyToken).toHaveBeenCalledTimes(1);
 		});
 
-		it('rejects when consumeJti is false and no expectedAudience is supplied', async () => {
-			// The overload signatures require `expectedAudience` whenever `consumeJti: false`
-			// is passed; bypass that compile-time guard to exercise the runtime fail-closed
-			// check, since it also protects callers whose audience is only optional at runtime
-			// (e.g. sourced from config).
-			const verifyTokenUnsafe = service.verifyToken.bind(service) as (
-				token: string,
-				options?: Record<string, unknown>,
-			) => Promise<unknown>;
+		it('should also verify and resolve an actor_token when present', async () => {
+			const actorClaims = { ...validClaims, sub: 'actor-user-1', jti: 'actor-jti-1' };
+			const actorUser = mock<User>({ id: 'actor-id', role: GLOBAL_MEMBER_ROLE });
+			externalTokenVerifierService.verifyToken
+				.mockResolvedValueOnce({ claims: validClaims, resolvedKey })
+				.mockResolvedValueOnce({ claims: actorClaims, resolvedKey });
+			identityResolutionService.resolve
+				.mockResolvedValueOnce(actorUser)
+				.mockResolvedValueOnce(mockUser);
+			config.maxTokenTtl = 900;
+			jwtService.sign.mockReturnValue('issued.jwt.token');
 
-			await expect(verifyTokenUnsafe('token', { consumeJti: false })).rejects.toThrow(
-				TokenExchangeAuthError,
-			);
-		});
-
-		it('rejects when the token audience does not match the supplied expectedAudience', async () => {
-			mockValidToken();
-			vi.spyOn(jwt, 'verify').mockImplementation(() => {
-				throw new Error('jwt audience invalid');
+			const result = await service.exchange({
+				grant_type: 'urn:ietf:params:oauth:grant-type:token-exchange',
+				subject_token: 'subject-token',
+				actor_token: 'actor-token',
 			});
 
-			await expect(
-				service.verifyToken('token', {
-					expectedAudience: 'unexpected-audience',
-					consumeJti: false,
-				}),
-			).rejects.toThrow(TokenExchangeAuthError);
-			expect(jwt.verify).toHaveBeenCalledWith(
-				'token',
-				resolvedKey.key,
-				expect.objectContaining({ audience: 'unexpected-audience' }),
-			);
-		});
-
-		it('accepts a set of acceptable audiences, passing them through as a tuple to jwt.verify', async () => {
-			mockValidToken();
-
-			await service.verifyToken('token', {
-				expectedAudience: ['aud-a', 'aud-b'],
-				consumeJti: false,
-			});
-
-			expect(jwt.verify).toHaveBeenCalledWith(
-				'token',
-				resolvedKey.key,
-				expect.objectContaining({ audience: ['aud-a', 'aud-b'] }),
-			);
-		});
-
-		it('rejects when consumeJti is false and expectedAudience is an empty array', async () => {
-			const verifyTokenUnsafe = service.verifyToken.bind(service) as (
-				token: string,
-				options?: Record<string, unknown>,
-			) => Promise<unknown>;
-
-			await expect(
-				verifyTokenUnsafe('token', { expectedAudience: [], consumeJti: false }),
-			).rejects.toThrow(TokenExchangeAuthError);
-		});
-
-		it('verifies the same token twice when consumeJti is false', async () => {
-			mockValidToken();
-
-			await service.verifyToken('token', {
-				expectedAudience: 'n8n',
-				consumeJti: false,
-			});
-			await service.verifyToken('token', {
-				expectedAudience: 'n8n',
-				consumeJti: false,
-			});
-
-			expect(jtiStore.consume).not.toHaveBeenCalled();
-		});
-
-		it('accepts a token without jti when requireJti is false', async () => {
-			const { jti: _, ...claimsWithoutJti } = validClaims;
-			vi.spyOn(jwt, 'decode').mockReturnValue({
-				header: { alg: 'RS256', kid: 'test-kid' },
-				payload: claimsWithoutJti,
-				signature: 'sig',
-			} as unknown as ReturnType<typeof jwt.decode>);
-			vi.spyOn(jwt, 'verify').mockReturnValue(
-				claimsWithoutJti as unknown as ReturnType<typeof jwt.verify>,
-			);
-			trustedKeyStore.getByKidAndIss.mockResolvedValue(resolvedKey);
-
-			const result = await service.verifyToken('token', {
-				expectedAudience: 'n8n',
-				consumeJti: false,
-				requireJti: false,
-			});
-
-			expect(result.claims.jti).toBeUndefined();
-			expect(jtiStore.consume).not.toHaveBeenCalled();
-		});
-
-		describe('subjectClaim override', () => {
-			it('leaves sub unchanged when subjectClaim is the default', async () => {
-				mockValidToken();
-				jtiStore.consume.mockResolvedValue(true);
-
-				const result = await service.verifyToken('valid-token');
-
-				expect(result.claims.sub).toBe(validClaims.sub);
-			});
-
-			it('substitutes the configured claim for sub when present', async () => {
-				const payload = { ...validClaims, uid: 'stable-okta-id' };
-				vi.spyOn(jwt, 'decode').mockReturnValue({
-					header: { alg: 'RS256', kid: 'test-kid' },
-					payload,
-					signature: 'sig',
-				} as unknown as ReturnType<typeof jwt.decode>);
-				vi.spyOn(jwt, 'verify').mockReturnValue(
-					payload as unknown as ReturnType<typeof jwt.verify>,
-				);
-				trustedKeyStore.getByKidAndIss.mockResolvedValue({ ...resolvedKey, subjectClaim: 'uid' });
-				jtiStore.consume.mockResolvedValue(true);
-
-				const result = await service.verifyToken('valid-token');
-
-				expect(result.claims.sub).toBe('stable-okta-id');
-			});
-
-			it('rejects when the configured claim is missing from the payload', async () => {
-				mockValidToken();
-				trustedKeyStore.getByKidAndIss.mockResolvedValue({ ...resolvedKey, subjectClaim: 'uid' });
-
-				await expect(service.verifyToken('valid-token')).rejects.toThrow(TokenExchangeAuthError);
-			});
-
-			it('rejects when the configured claim is present but not a string', async () => {
-				const payload = { ...validClaims, uid: 12345 };
-				vi.spyOn(jwt, 'decode').mockReturnValue({
-					header: { alg: 'RS256', kid: 'test-kid' },
-					payload,
-					signature: 'sig',
-				} as unknown as ReturnType<typeof jwt.decode>);
-				vi.spyOn(jwt, 'verify').mockReturnValue(
-					payload as unknown as ReturnType<typeof jwt.verify>,
-				);
-				trustedKeyStore.getByKidAndIss.mockResolvedValue({ ...resolvedKey, subjectClaim: 'uid' });
-
-				await expect(service.verifyToken('valid-token')).rejects.toThrow(TokenExchangeAuthError);
-			});
-		});
-	});
-
-	describe('verifyExternalToken', () => {
-		it('verifies a Keycloak-issued token end-to-end and returns attested claims', async () => {
-			const { publicKey, privateKey } = generateKeyPairSync('rsa', { modulusLength: 2048 });
-			const keycloakClaims = {
-				sub: 'a1b2c3d4-keycloak-subject',
-				iss: 'https://keycloak.example.com/realms/n8n',
-				aud: 'n8n-resource-server',
-				iat: now,
-				exp: now + 300,
-				jti: 'keycloak-token-id',
-				email: 'user@example.com',
-				email_verified: true,
-				given_name: 'Ada',
-				family_name: 'Lovelace',
-			};
-			const token = jwt.sign(keycloakClaims, privateKey, {
-				algorithm: 'RS256',
-				keyid: 'keycloak-kid',
-			});
-
-			trustedKeyStore.getByKidAndIss.mockResolvedValue({
-				sourceId: 'keycloak-realm',
-				kid: 'keycloak-kid',
-				algorithms: ['RS256'],
-				key: publicKey,
-				issuer: keycloakClaims.iss,
-				requireVerifiedEmail: false,
-				subjectClaim: 'sub',
-			});
-
-			const result = await service.verifyExternalToken(token, keycloakClaims.aud);
-
-			expect(result).toEqual({
-				claim: {
-					sourceId: 'keycloak-realm',
-					issuer: keycloakClaims.iss,
-					subject: keycloakClaims.sub,
-					audience: keycloakClaims.aud,
-					attributes: {
-						email: keycloakClaims.email,
-						email_verified: true,
-						given_name: 'Ada',
-						family_name: 'Lovelace',
-					},
-					expiresAt: new Date(keycloakClaims.exp * 1000),
-				},
-			});
-			expect(jtiStore.consume).not.toHaveBeenCalled();
-		});
-
-		it('returns a typed context instead of throwing when verification fails', async () => {
-			await expect(service.verifyExternalToken('garbage-token', 'n8n')).resolves.toEqual({
-				claim: null,
-				context: { reason: 'invalid_token', errorDetails: expect.any(String) },
-			});
-		});
-
-		it('substitutes a configured subjectClaim for sub end-to-end (Okta-style divergence)', async () => {
-			const { publicKey, privateKey } = generateKeyPairSync('rsa', { modulusLength: 2048 });
-			const oktaClaims = {
-				sub: 'alice.example@acme.com',
-				uid: 'okta-immutable-id-42',
-				iss: 'https://acme.okta.com/oauth2/aus1abc',
-				aud: 'n8n-resource-server',
-				iat: now,
-				exp: now + 300,
-				jti: 'okta-token-id',
-			};
-			const token = jwt.sign(oktaClaims, privateKey, { algorithm: 'RS256', keyid: 'okta-kid' });
-
-			trustedKeyStore.getByKidAndIss.mockResolvedValue({
-				sourceId: 'okta-custom-as',
-				kid: 'okta-kid',
-				algorithms: ['RS256'],
-				key: publicKey,
-				issuer: oktaClaims.iss,
-				requireVerifiedEmail: false,
-				subjectClaim: 'uid',
-			});
-
-			const result = await service.verifyExternalToken(token, oktaClaims.aud);
-
-			expect(result).toMatchObject({ claim: { subject: oktaClaims.uid } });
+			expect(result.actor).toBe('actor-user-1');
+			expect(result.actorUserId).toBe('actor-id');
+			expect(externalTokenVerifierService.verifyToken).toHaveBeenCalledTimes(2);
 		});
 	});
 });
